Add auth mechanizm

+++ BL/DTOs/BookDto.cs
@@ -12,7 +12,7 @@ namespace BL.DTOs
         public string Genre { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
-        public int Price { get; set; }
+        public float Price { get; set; }
         public IEnumerable<BookRevisionDto> BookRevisions { get; set; }
     }
 }
